pyproject.toml
[project]
name = "python-odata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adal>=1.2.7",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.0.1",
]
